requirements.txt
requests>=2.25.0
beautifulsoup4>=4.9.0
pandas>=1.0.0
openai>=1.0.0
pyyaml>=5.4.0
psutil>=5.8.0
kaggle>=1.5.0
python-dotenv>=0.19.0